refactor(contracts): 更新VRFCoordinator参数和Raffle合约
- 修改VRFCoordinatorMock的gasPriceLink参数为2e18
- 将FundSubscription合约中的FUND_AMOUNT类型从uint96改为uint256
- 在Raffle合约中添加AutomationCompatibleInterface接口
- 更新测试文件中的逻辑和注释

+++ script/Interactions.s.sol
@@ -47,7 +47,7 @@ contract CreateSubscription is Script {
 
 /**为订阅提供资金 */
 contract FundSubscription is Script {
-    uint96 public constant FUND_AMOUNT = 3 ether;
+    uint256 public constant FUND_AMOUNT = 3 ether;
 
     function fundSubscriptionUsingConfig() public {
         // HelperConfig helperConfig = new HelperConfig();
